pyproject.toml
[project]
name = "jay-trading"
version = "0.1.0"
description = "AI-executed paper trading system driven by congressional and insider disclosure signals (Alpaca paper only)."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Jay Nayee", email = "jvn3@njit.edu" }
]
dependencies = [
    "alembic>=1.18.4",
    "alpaca-py>=0.43.2",
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pandas>=3.0.2",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "tenacity>=9.1.4",
]

[project.scripts]
jay-trading = "jay_trading.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: hits real network (skipped in CI by default)",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
    "vcrpy>=8.1.1",
]
